Refactor `interfaces` to use composition over inheritance
For instance the `Reporter` class becomes a module that can be included
in any class, without enforcing the class to inherit from the old
abstract class.

Basically this implies that any class can act as a 'reporter', yet it
does not need actualy to inherits from it, if implements the necessary
interface.

Affected interfaces: Matcher, Order, Output, Reporter, Runner